+++ flutter_todo_app/.metadata
@@ -15,22 +15,7 @@ migration:
     - platform: root
       create_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
       base_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-    - platform: android
-      create_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-      base_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-    - platform: ios
-      create_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-      base_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-    - platform: linux
-      create_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-      base_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-    - platform: macos
-      create_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-      base_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
     - platform: web
-      create_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-      base_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
-    - platform: windows
       create_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
       base_revision: f468f3366c26a5092eb964a230ce7892fda8f2f8
